unverified-contract-py/requirements.txt
forta_bot_sdk>=0.2.3
setuptools>=61.3.1
pyevmasm>=0.2.3
rlp>=2.0.1
hexbytes>=0.2.2
bot-alert-rate>=0.0.4
python-dotenv>=1.0.1